Add schema tables for book adn label and organize files
