+++ books/JSON_files/books.json
@@ -1,1 +1,1 @@
-[]
+[{"id":516,"publisher":"mus","cover_state":"good","publish_date":"2020-05-09","archived":false,"label":"AAA"},{"id":615,"publisher":"ter","cover_state":"bad","publish_date":"2001-08-06","archived":true,"label":"TTT"},{"id":317,"publisher":"fff","cover_state":"good","publish_date":"2003-09-12","archived":true,"label":"RER"},{"id":594,"publisher":"ioio","cover_state":"good","publish_date":"2015-07-12","archived":false,"label":"KKLKL"}]
