+++ books/JSON_files/labels.json
@@ -1,1 +1,1 @@
-[]
+[{"title":"AAA","color":"ZZZZ"},{"title":"TTT","color":"GGG"},{"title":"RER","color":"FDFD"},{"title":"KKLKL","color":"MÙMÙM"}]
